chore(update): update styling files for both templates
