+++ src/components/pages/MahomesPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { MahomesTemplate } from 'src/templates/MahomesTemplate';
-import data from '../../example-json/john_smith.json';
+import data from '../../example-json/patrick_dohn.json';
 
 export const MahomesPage = () => {
   return (
