+++ src/components/templates/FolesTemplate/Content/DetailHeading.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { Text, View } from 'react-pdf';
 
-import { styles } from './content-styles';
+import { styles } from '../Styles';
 
 export const DetailHeading = ({
   company,
